\name{easyLook}
\alias{easyLook}
\title{Summarize clipper output.}
\description{Summarization of the result for a quick look of clipper function.}
\usage{easyLook(clipped)}
\arguments{
  \item{clipped}{the output of either clipper o easyClip.}
}

\value{Nice formatted output of clipper function.}

\references{
  Martini P, Sales G, Massa MS, Chiogna M, Romualdi C. Along signal
  paths: an empirical gene set approach exploiting pathway
  topology. NAR. 2012 Sep.

  Massa MS, Chiogna M, Romualdi C. Gene set analysis exploiting the
  topology of a pathway. BMC System Biol. 2010 Sep 1;4:121.
}
  
