% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-through-poly.R
\name{path_through_poly}
\alias{path_through_poly}
\title{path_through_poly}
\usage{
path_through_poly(path, polys)
}
\arguments{
\item{path}{A \code{data.frame} with \code{x} and \code{y} columns denoting sequential
coordinates of a path.}

\item{polys}{A list of \code{data.frame} objects, each of which has coordinates
labelled \code{x} and \code{y} tracing a closed polygon.}
}
\value{
Single value quantifying the total relative length of the input path
that passes through the polygons
}
\description{
Relative length of path that passes through polygons
}
